refactor: moved interfaces to domain layer
