+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/remote/gateway/LocationGateway.kt
@@ -1,8 +1,9 @@
-package org.thechance.common.data.remote.gateway.location_gateway
+package org.thechance.common.data.remote.gateway
 
 import org.thechance.common.domain.entity.Location
+import org.thechance.common.domain.getway.ILocationGateway
 
-class LocationGateway:ILocationGateway {
+class LocationGateway: ILocationGateway {
     override suspend fun getCurrentLocation(): Location {
         TODO("Not yet implemented")
     }
