+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/remote/gateway/TaxiGateway.kt
@@ -1,10 +1,11 @@
-package org.thechance.common.data.remote.gateway.taxi_gateway
+package org.thechance.common.data.remote.gateway
 
 import org.thechance.common.domain.entity.DataWrapper
 import org.thechance.common.domain.entity.NewTaxiInfo
 import org.thechance.common.domain.entity.Taxi
+import org.thechance.common.domain.getway.ITaxiGateway
 
-class TaxiGateway:ITaxiGateway {
+class TaxiGateway: ITaxiGateway {
     override suspend fun getTaxis(): DataWrapper<Taxi> {
         TODO("Not yet implemented")
     }
