+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/domain/getway/ITaxiGateway.kt
@@ -1,11 +1,11 @@
-package org.thechance.common.data.remote.gateway.taxi_gateway
+package org.thechance.common.domain.getway
 
-import org.thechance.common.domain.entity.AddTaxi
 import org.thechance.common.domain.entity.DataWrapper
+import org.thechance.common.domain.entity.NewTaxiInfo
 import org.thechance.common.domain.entity.Taxi
 
 interface ITaxiGateway {
     suspend fun getTaxis(): DataWrapper<Taxi>
-    suspend fun createTaxi(taxi: AddTaxi): Taxi
+    suspend fun createTaxi(taxi: NewTaxiInfo): Taxi
     suspend fun searchTaxiByDriverUsername(driverUsername: String): DataWrapper<Taxi>
 }
